refactor(boundaries): extract activity-feed text helpers to core protocol (refs #1363)
Core modules (cockpit_activity) had to lazy-import format_relative_time
and render_entry_detail from the activity_feed plugin's feed_panel,
making the plugin effectively non-optional. Moved both pure helpers plus
an ActivityFeedEntry Protocol to a new pollypm.activity_feed_protocol
module that core and the plugin can both depend on. The plugin re-exports
the names from feed_panel so existing imports + tests keep working.
