+++ src/pollypm/plugins_builtin/activity_feed/cockpit/feed_panel.py
@@ -21,7 +21,6 @@ projection only runs when something changed.
 
 from __future__ import annotations
 
-import json
 import logging
 from dataclasses import dataclass, replace
 from datetime import UTC, datetime, timedelta
@@ -167,38 +166,11 @@ _SEVERITY_COLOURS = {
 }
 
 
-def format_relative_time(
-    timestamp: str, *, now: datetime | None = None,
-) -> str:
-    """Return a relative-time label (``"3m ago"``, ``"2h ago"``).
-
-    ``timestamp`` is an ISO-8601 string (the projector always produces
-    these). Unparseable values fall back to the raw string. ``now`` is
-    injectable for deterministic tests.
-    """
-    if not timestamp:
-        return "—"
-    try:
-        when = datetime.fromisoformat(timestamp)
-    except ValueError:
-        return timestamp
-    if when.tzinfo is None:
-        when = when.replace(tzinfo=UTC)
-    anchor = now or datetime.now(UTC)
-    delta = anchor - when
-    seconds = int(delta.total_seconds())
-    if seconds < 0:
-        return "just now"
-    if seconds < 60:
-        return f"{seconds}s ago"
-    minutes = seconds // 60
-    if minutes < 60:
-        return f"{minutes}m ago"
-    hours = minutes // 60
-    if hours < 24:
-        return f"{hours}h ago"
-    days = hours // 24
-    return f"{days}d ago"
+# Re-exported from the core shared module so callers in ``pollypm.cockpit*``
+# (and other peer plugins) can use the helper without importing through
+# ``plugins_builtin`` (#1363). The plugin keeps the name on its own surface
+# for back-compat with existing imports + monkeypatch paths.
+from pollypm.activity_feed_protocol import format_relative_time  # noqa: F401, E402
 
 
 def _project_label(entry: FeedEntry) -> str:
@@ -286,57 +258,10 @@ def render_entries_as_text(entries: Iterable[FeedEntry]) -> str:
     return "\n".join(rows)
 
 
-def render_entry_detail(entry: FeedEntry, *, now: datetime | None = None) -> str:
-    """Render the per-entry detail view as plain text (lf04).
-
-    Shows absolute + relative timestamps, project / actor / verb,
-    severity, a pretty-printed payload JSON, and suggested follow-up
-    navigation (task / session links) inferred from the payload.
-    """
-    lines: list[str] = []
-    abs_ts = entry.timestamp
-    rel_ts = format_relative_time(entry.timestamp, now=now)
-    lines.append(f"Activity entry · {entry.id}")
-    lines.append("")
-    lines.append(f"When: {abs_ts}  ({rel_ts})")
-    lines.append(f"Kind: {entry.kind}")
-    if entry.project:
-        lines.append(f"Project: {entry.project}")
-    lines.append(f"Actor: {entry.actor}")
-    if entry.subject:
-        lines.append(f"Subject: {entry.subject}")
-    lines.append(f"Verb: {entry.verb}")
-    lines.append(f"Severity: {entry.severity}")
-    if entry.pinned:
-        lines.append("Pinned: yes")
-    lines.append("")
-    lines.append("Summary:")
-    lines.append(f"  {entry.summary}")
-    # Navigation hints — task / session links inferred from the payload.
-    task_project = entry.payload.get("task_project")
-    task_number = entry.payload.get("task_number")
-    if task_project and task_number is not None:
-        lines.append("")
-        lines.append(
-            f"Related task: project:{task_project}:task:{task_number}  "
-            f"(use the rail to open)"
-        )
-    elif entry.source == "work_transitions" and entry.subject:
-        lines.append("")
-        lines.append(f"Related task: {entry.subject}")
-    if entry.actor and entry.source == "events":
-        lines.append("")
-        lines.append(f"Related session: {entry.actor}")
-    # Payload dump (pretty-printed).
-    lines.append("")
-    lines.append("Payload:")
-    try:
-        rendered = json.dumps(entry.payload, indent=2, sort_keys=True, default=str)
-    except (TypeError, ValueError):
-        rendered = repr(entry.payload)
-    for line in rendered.splitlines():
-        lines.append(f"  {line}")
-    return "\n".join(lines)
+# Re-exported from the core shared module (#1363). The plugin's ``FeedEntry``
+# dataclass satisfies ``ActivityFeedEntry`` structurally, so the renderer
+# accepts the existing concrete type unchanged.
+from pollypm.activity_feed_protocol import render_entry_detail  # noqa: F401, E402
 
 
 def render_activity_feed_text(config: Any, *, limit: int = 50) -> str:
